// src/vega.schema.ts
export interface VgSignalRef {
  signal: string;
}
// ...
export interface VgFormulaTransform {
  type: 'formula';
  as: string;
  expr: string;
}
// ...
export interface VgImputeTransform {
  type: 'impute';
  groupby?: string[];
  field: string;
  key: string;
  keyvals?: string[];
  method?: 'value' | 'median' | 'max' | 'min' | 'mean';
  value?: any;
}
// ...
export type VgComparatorOrder = 'ascending' | 'descending';

export interface VgComparator {
  field?: string | string[];
  order?: VgComparatorOrder | VgComparatorOrder[];
}

export interface VgWindowTransform {
  type: 'window';
  params?: Number[];
  as?: string[];
  ops?: (AggregateOp | WindowOnlyOp)[];
  fields?: string[];
  frame?: Number[];
  ignorePeers?: Boolean;
  groupby?: string[];
  sort?: VgComparator;
}